+++ src/imio/smartweb/core/webcomponents/src/components/Search/EventsResult/EventsResult.jsx
@@ -12,7 +12,7 @@ const EventsResult = (props) => {
             headers: {
                 Accept: "application/json",
             },
-            params: props && props.urlParams,
+            params: (props.urlParams.SearchableText || props.urlParams.iam || props.urlParams.topics) ? props.urlParams : {},
         },
         [props]
     );